apps/api/pyproject.toml
[project]
name = "flow-cytometry-api"
version = "0.1.0"
description = "FastAPI backend for the flow cytometry automation platform."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13",
  "fastapi>=0.110",
  "httpx>=0.27",
  "minio>=7.2",
  "psycopg[binary]>=3.1",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0",
  "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
